# Initial commit of the ProcessDependencyGraph Tooling
This tooling was spun out of MsBuildProjectReferenceDependencyGraph, specifically the undocumented switch -TargetProject. Historically that switch modified the DOT Graph produced by this tooling to highlight a specific project on the graph and then identify both the dependencies required to build the highlighted project and the projects which would be affected by a change to the project.

This is additional post processing on the dependency graph which increased the complexity of MsBuildProjectReferenceDependencyGraph without being related to its primary purpose. Therefore this logic was spun out into a new tool (ProcessDependencyGraph) that can work on the DOT Graph generated by MsBuildProjectReferenceDependencyGraph.

However the exact behavior cannot be replaced in the new tooling in a single step as that tooling has split those graphs out into two separate tasks. For my purposes I generally wanted to answer one question or the other, but rarely (if ever both) at the same time. If there is a pull request to restore the functionality as it previously existed we can address it there.

A Comprehensive README.md has been created for ProcessDependencyGraph that covers the usage of this tool as well as the new features introduced by it.